Removed the mod locking code
It was decided it's not a feature worth the time to support. At least, not at this time and in this form. It's not worth building a UI for it.

It was originally created in State Alchemist as a way to prevent certain mods from changing when running random. However, Simple Mod Alchemist can now change things at random individually or in groups, somewhat defeating the importance of this feature.